Add a missing random triplet sample function
The planar triplet sample function samples three lines
where their directions vector are in a plane.
This yields to orthogonals which are all parallel.
If these lines also intersect then the lines are really in a plane.

The plane triplet is needed as the triple product
of the direction vectors yields to zero
which could be an annoying special case.