Add shared object target
Now, a shared object can be created.
It was not intended to build a shared object and it uses its own PIC
based objects such that the built is completely independant of the
static build.
To spare compilation time it is separated and must be triggered
manually meaning the build is not part of the all target.

In general, this commit cleans the CMakeLists.txt a bit